+++ src/voyage_engine/assets/AssetManager.java
@@ -19,15 +19,15 @@ import voyage_engine.assets.texture.Texture;
 
 public class AssetManager {
 	private static Manifest manifest;
-	private static HashMap<Long, Asset> assetMap;
+	private static HashMap<Integer, Asset> assetMap;
 
-	private static LinkedList<Long> unloadIdList;
+	private static LinkedList<Integer> unloadIdList;
 
 	public static void initialize(boolean rebaseManifest) {
 		// Start spool and use the default initializer to auto detect thread count.
 		Spool.initialize();
-		assetMap = new HashMap<Long, Asset>();
-		unloadIdList = new LinkedList<Long>();
+		assetMap = new HashMap<Integer, Asset>();
+		unloadIdList = new LinkedList<Integer>();
 
 		System.out.println("[manifest]: loading manifest...");
 		if (!rebaseManifest) { // if we are not rebasing, just attempt to load the manifest.
@@ -67,7 +67,7 @@ public class AssetManager {
 	// using
 	// multi-threaded Loading.
 	public static Texture getTexture(AssetCache cache, String filename, boolean filter, boolean mipmap) {
-		long id = manifest.getID(filename);
+		int id = manifest.getID(filename);
 		// if a cache was provided the id will be include in the cache list to free when
 		// cache is no longer needed.
 		if (cache != null) {
@@ -90,7 +90,7 @@ public class AssetManager {
 	// returns the requested font by filename if it can be found in the content map
 	// or returns a new pointer to the font and loads the font immediately.
 	public static Font getFont(AssetCache cache, String filename, int oversampling, boolean filter) {
-		Long id = manifest.getID(filename);
+		int id = manifest.getID(filename);
 		// if a cache was provided the id will be include in the cache list to free when
 		// cache is no longer needed.
 		if (cache != null) {
@@ -114,7 +114,7 @@ public class AssetManager {
 
 	// returns the requested shader file
 	public static Shader getShader(AssetCache cache, String filename) {
-		long id = manifest.getID(filename);
+		int id = manifest.getID(filename);
 		// if a cache was provided the id will be include in the cache list to free when
 		// cache is no longer needed.
 		if (cache != null) {
@@ -213,7 +213,7 @@ public class AssetManager {
 		}
 	}
 
-	public static void release(Long id) {
+	public static void release(int id) {
 		Asset asset = assetMap.get(id);
 		release(asset);
 	}
@@ -231,7 +231,7 @@ public class AssetManager {
 
 	private static void releaseUnreferencedAssets() {
 		// delete all the id's in the unload list.
-		for (Long id : unloadIdList) {
+		for (int id : unloadIdList) {
 			assetMap.remove(id);
 		}
 		unloadIdList.clear();
